Валидация формы заказа, создание кастомного правила валидации на мобильный телефон

+++ database/migrations/2024_05_10_103046_create_delivery_types_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
 
             $table->string('title');
 
-            $table->unsignedBigInteger('price')
+            $table->unsignedInteger('price')
                 ->default(0);
 
             $table->boolean('with_address')
